requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
PyJWT==2.9.0
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
